requirements.txt
djangorestframework
django-filter
pyjwt==1.7.1
djangorestframework-simplejwt[crypto]
django-storages
boto3
python-dotenv
markdown
pillow
django-mailjet
django-cors-headers
django-resized
transformers[torch]
channels[daphne]
django-unfold
python-decouple
kaleido
plotly
pandas
dj-database-url
